[Types] type.h 추가
enum conflict 방지를 위하여 literal 관리 클레스 추가
