+++ EMS/EMS/Dbms.h
@@ -2,26 +2,11 @@
 
 #include "Employee.h"
 #include "DataBase.h"
-#include "../EMS/Dbms.h"
 #include <list>
 
 using namespace std;
 
-enum class Column {
-	EMPLOYEE_NUM,
-	NAME,
-	FIRST_NAME,
-	LAST_NAME,
-	CL,
-	PHONE,
-	PHONE_MID,
-	PHONE_LAST,
-	BIRTHDAY,
-	BIRTHDAY_YEAR,
-	BIRTHDAY_MONTH,
-	BIRTHDAY_DAY,
-	CERTI
-};
+
 
 class IDbms {
 public:
